pyproject.toml
[project]

name = "f1-telemetry-manager"

version = "0.1.0"

description = "F1 telemetry Dashboard"

readme = "README.md"
authors = [
    {name = "Víctor Vega Sobral", email = "https://www.linkedin.com/in/victorvegasobral/"},
    {name = "Santiago Souto Ortega", email = "https://www.linkedin.com/in/santiagosoutoortega/"}

]

requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "supabase",
    "python-dotenv",
    "pydantic[email]",
    "fastf1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0"
]